moved list_versions to a versioning_helpers.py
